# Spring boot/SB-Customer-Rest-App/Customer-Rest-Api-Documentation.json
{
	"swagger": "2.0",
	"info": {
		"description": "Spring Boot REST API for Online Store",
		"version": "1.0",
		"title": "Spring Boot REST API",
		"termsOfService": "Terms of service",
		"contact": {
			"name": "Naresh I Technologies",
			"url": "www.nareshit.com",
			"email": "info@nareshit.com"
		},
		"license": {
			"name": "Apache License Version 2.0",
			"url": "https://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:9090",
	"basePath": "/",
	"tags": [
		{
			"name": "customer-rest-controller",
			"description": "Customer Rest Controller"
		}
	],
	"paths": {
		"/add": {
			"post": {
				"tags": [
					"customer-rest-controller"
				],
				"summary": "This operation is used to save customer details",
				"operationId": "addCustomerUsingPOST",
				"consumes": [
					"application/xml",
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "c",
						"description": "This represents Customer Data",
						"required": false,
						"schema": {
							"$ref": "#/definitions/This represents Model class for Customer Data"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created",
						"schema": {
							"type": "string"
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Server Problem"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/get/{id}": {
			"get": {
				"tags": [
					"customer-rest-controller"
				],
				"summary": "This Operations is used to provide Customer Details Based on Customer ID",
				"operationId": "getCustomerUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/xml",
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "This represents Customer Id",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/This represents Model class for Customer Data"
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Server Problem"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"This represents Model class for Customer Data": {
			"type": "object",
			"properties": {
				"customerEmail": {
					"type": "string",
					"description": "This Customer Email"
				},
				"customerId": {
					"type": "integer",
					"format": "int32",
					"description": "This is Customer ID"
				},
				"customerName": {
					"type": "string",
					"description": "This is Customer Name"
				}
			}
		}
	}
}